GS-108 feat(PlayerCharacter): Spaceship movement
Spaceship lateral movements have been added. It is now possible to control the spaceship with a tap on the screen.

The spaceship will move to the opposite side of the screen (using the screen center as a reference) whenever the player taps the screen.

It is possible to hold the touch to charge a larger propel released when fully charged or when the touch input is removed.