GS-111 feat(GamePhysics): Gravity fields
Lateral gravitational fields were created. Now the spaceship is pulled sideways to the sides of the screen.

The starting side to which the spaceship will be pulled is chosen at random.

By touching the sides of the screen, the spaceship is moved back to the center of the screen. This mechanic was just put as a placeholder to make tests more viable.