GS-103 feat(obstacles): Dodgeable Obstacle Element
The Dodgeable Obstacle now moves vertically across the screen.
The movement speed can be changed by changing the ObstacleVel variable in the BP_DodgdableObstacle blueprint.

The player now collides with the Dodgeable Obstacle, causing the PlayerCharacter to move back to the placeholder.

The Dodgeable Obstacle has a spawn cycle by varying its position. Spawn speed can be changed in the SpawnVel variable in the BP_ElementsSpawner blueprint.