GS-112 feat(PlayerCharacter): Player Shoot
The spacecraft now shoots a projectile continuously in front of it.
The shot follows a linear trajectory, destroying itself if it collides with an Obstacle Tag object or if it leaves the screen.

To change the speed at which the shots occur, change the ShootSpawnVel variable in the BP_CharacterSpaceship blueprint.