fix(PlayerCharacter): Spaceship movement
The spacecraft motion was having trouble handling the propulsion charge and propelling the spacecraft variably according to touch input time.

The issue has been resolved, and the spaceship now correctly charges the boost whether the player holds the screen or not.

Now, if the player holds the touch input until the charge is completely filled, the charge will automatically release as intended.