GS-30/GS-32 feat(Obstacles): Destructible Obstacle
The Destructible Obstacle was implemented as the BP_DestructibleObstacle blueprint actor.

The Destructible Obstacle had its speed implemented. The speed can be changed by modifying the ObstacleVel variable of the BP_DestructibleObstacle blueprint.

The Destructible Obstacle gap size can be changed by modifying the ObstacleGap variable of the BP_DestructibleObstacle blueprint.

The collision of the player's spaceship with the Destructible Obstacle has been implemented.