GS-180 feat(GamePlay): Arrows Referencing
This change adds arrows referencing the direction of the propel.

The arrows change as the spacecraft changes sides on the screen, indicating which direction the spacecraft will move in an eventual propel.

The code was implemented in the ReferenceArrows function in the BP_CharacterSpaceship blueprint and is called continuously by the Event Tick