GS-182 fix(GameElements): X Axis Movement Bug
This change resolves the X-axis movement issue that was happening erroneously in some elements.

The change replaces the Move Component To action to a Timeline that updates the Y-axis movement using Set Relative Location. In this way, the possibility of any unwanted interaction with the elements' Transform information is eliminated.